backend/pyproject.toml
[project]
name = "aristotle-backend"
version = "0.1.0"
description = "Aristotle — gamified course companion backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "sqlmodel>=0.0.22",
    "python-multipart>=0.0.12",
    "httpx>=0.27.0",
    "anthropic>=0.39.0",
    "openai>=1.54.0",
    "chromadb>=0.5.15",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-chroma>=0.1.0",
    "langchain-anthropic>=0.2.0",
    "langgraph>=0.2.30",
    "pypdf>=5.1.0",
    "pdfplumber>=0.11.0",
    "pillow>=11.0.0",
    "python-dotenv>=1.0.1",
    "tenacity>=9.0.0",
    "langchain-community>=0.4.1",
    "langchain-text-splitters>=1.1.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
